src/requirements.txt
# Core Frameworks
fastapi==0.115.12
uvicorn==0.34.3
google-adk==1.3.0

# AI & Google Cloud
google-generativeai
google-cloud-aiplatform==1.97.0
google-auth==2.39.0
google-cloud-secret-manager==2.24.0
google-cloud-logging==3.11.4

# Database - PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data & Web
requests==2.32.4
beautifulsoup4==4.13.4
pandas==2.3.0
yfinance==0.2.33
sec-api==1.0.17

# Utilities
python-dotenv==1.1.0
python-dateutil==2.9.0.post0
pydantic==2.11.7
